--- api/rag_chatbot/docker-compose.yml
@@ -1,11 +1,0 @@
-version: "3.9"
-
-services:
-  fastapi-app:
-    build: .
-    ports:
-      - "8000:8000"
-    volumes:
-      - .:/app
-    environment:
-      - GROQ_API_KEY=${GROQ_API_KEY}
