--- api/rag_chatbot/requirements.txt
@@ -1,10 +1,0 @@
-fastapi
-uvicorn[standard]
-python-dotenv
-langchain
-langchain-community
-langchain-groq
-sentence-transformers
-faiss-cpu
-uvicorn
-fastapi
