+++ api/rag_chatbot/utils.py
@@ -1,28 +1,49 @@
-from fastapi import UploadFile, File, Form, APIRouter
-from fastapi.responses import JSONResponse
-from api.rag_chatbot.helper import RAGPipeline
-import tempfile
 import os
+from langchain_community.document_loaders import PyPDFLoader
+from langchain.text_splitter import RecursiveCharacterTextSplitter
+from langchain_community.vectorstores import FAISS
+from langchain_groq import ChatGroq
+from langchain.chains import RetrievalQA
+from dotenv import load_dotenv
+from langchain_community.embeddings import HuggingFaceBgeEmbeddings
 
-router = APIRouter()
-pipeline = RAGPipeline()
+load_dotenv()  # load variables from .env into environment
 
-@router.post("")
-async def ask_question_from_pdf(file: UploadFile = File(...), question: str = Form(...)):
-    # Save the uploaded PDF temporarily
-    with tempfile.NamedTemporaryFile(delete=False, suffix=".pdf") as tmp:
-        tmp.write(await file.read())
-        tmp_path = tmp.name
+groq_api_key = os.getenv("GROQ_API_KEY")
 
-    try:
-        # Process PDF with the RAG pipeline
-        docs = pipeline.load_pdf(tmp_path)
-        chunks = pipeline.split_documents(docs)
-        vectorstore = pipeline.create_vectorstore(chunks)
-        llm = pipeline.get_llm()
-        response = pipeline.ask_question(vectorstore, question, llm)
-        return JSONResponse(content={"response": response['result']})
-    except Exception as e:
-        return JSONResponse(content={"error": str(e)}, status_code=500)
-    finally:
-        os.remove(tmp_path)  # Always clean up temp file
+
+class RAGPipeline:
+    def __init__(self, embedding_model="BAAI/bge-small-en", device="cpu"):
+        self.embedding_model = embedding_model
+        self.device = device
+        self.embeddings = self._load_embeddings()
+    
+    def _load_embeddings(self):
+        model_kwargs = {"device": self.device}
+        encode_kwargs = {"normalize_embeddings": True}
+        return HuggingFaceBgeEmbeddings(
+            model_name=self.embedding_model,
+            model_kwargs=model_kwargs,
+            encode_kwargs=encode_kwargs,
+        )
+    
+    def load_pdf(self, file_path):
+        loader = PyPDFLoader(file_path)
+        return loader.load()
+    
+    def split_documents(self, documents, chunk_size=500, chunk_overlap=100):
+        splitter = RecursiveCharacterTextSplitter(
+            chunk_size=chunk_size, chunk_overlap=chunk_overlap
+        )
+        return splitter.split_documents(documents)
+
+    def create_vectorstore(self, docs):
+        return FAISS.from_documents(docs, self.embeddings)
+    
+    def get_llm(self, model="meta-llama/llama-4-maverick-17b-128e-instruct"): 
+        return ChatGroq(model=model)
+    
+    def ask_question(self, vectorstore, question, llm):
+        retriever = vectorstore.as_retriever(search_type="similarity", search_kwargs={"k": 3})
+        qa = RetrievalQA.from_chain_type(llm=llm, retriever=retriever)
+        return qa.invoke(question)
