--- api/summarization/schemas.py
@@ -1,1 +1,0 @@
-#schemas.py file for schemas
