+++ api/summarization/views.py
@@ -1,32 +1,40 @@
-#views.py file for viewing api
-
-from fastapi import FastAPI, File, UploadFile
+from fastapi import APIRouter, UploadFile, File, HTTPException, Form
 from fastapi.responses import JSONResponse
-import shutil
+from api.summarization.utils import SummarizationPipeline
+import tempfile
 import os
-from api.service.utils import read_pdf,summarization
 
-app = FastAPI()
+router = APIRouter()
+summarizer = SummarizationPipeline()
+
+@router.post("/summarize-pdf")
+async def summarize_pdf(file: UploadFile = File(...)):
+    """Endpoint to summarize uploaded PDF files."""
+    with tempfile.NamedTemporaryFile(delete=False, suffix=".pdf") as tmp:
+        tmp.write(await file.read())
+        tmp_path = tmp.name
+
+    try:
+        text = summarizer.read_pdf(tmp_path)
+        summary = summarizer.summarize_document(text)
+        return JSONResponse(content={"summary": summary})
+    except Exception as e:
+        raise HTTPException(status_code=500, detail=f"PDF summarization failed: {str(e)}")
+    finally:
+        os.remove(tmp_path)
+
+@router.post("/summarize-youtube")
+async def summarize_youtube(url: str = Form(...)):
+    """Endpoint to summarize YouTube videos."""
+    try:
+        summary = summarizer.summarize_youtube(url)
+        return JSONResponse(content={"summary": summary})
+    except Exception as e:
+        raise HTTPException(status_code=500, detail=f"YouTube summarization failed: {str(e)}")
 
 
-@app.post("/extract-pdf-text/")
-async def extract_pdf_text(file: UploadFile = File(...)):
-    if not file.filename.endswith(".pdf"):
-        return JSONResponse(status_code=400, content={"error": "Only PDF files are allowed."})
 
-    # Save the uploaded file temporarily
-    temp_path = f"temp_{file.filename}"
-    with open(temp_path, "wb") as buffer:
-        shutil.copyfileobj(file.file, buffer)
 
-    # Read and extract text
-    text = read_pdf(temp_path)
 
-    # Delete temporary file
-    os.remove(temp_path)
 
-    summary = summarization(text)
 
-    return {"filename": file.filename, 
-            "summary": summary
-            }
